backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-dotenv
pytest
argon2-cffi
